+++ app/src/main/java/com/example/kotlinmusic/DisplayMusicsActivity.kt
@@ -1,4 +1,3 @@
-// DisplayMusicsActivity.kt
 package com.example.kotlinmusic
 
 import android.os.Bundle
@@ -8,22 +7,15 @@ import com.example.kotlinmusic.ui.theme.KotlinMusicTheme
 import com.google.android.exoplayer2.ExoPlayer
 
 class DisplayMusicsActivity : ComponentActivity() {
-    private lateinit var player: ExoPlayer
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        player = ExoPlayer.Builder(this).build()
-        val playlist = intent.getStringExtra("selected_playlist") ?: "Playlist 1"
+        val playlist = intent.getStringExtra("selected_playlist") ?: return
+        val player = ExoPlayer.Builder(this).build()
 
         setContent {
             KotlinMusicTheme {
                 DisplayMusics(player, playlist)
             }
         }
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        player.release()
     }
 }
