+++ app/src/main/java/com/example/kotlinmusic/MainActivity.kt
@@ -1,4 +1,3 @@
-// MainActivity.kt
 package com.example.kotlinmusic
 
 import android.content.Intent
@@ -9,8 +8,12 @@ import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material3.*
+import androidx.compose.material.*
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Home
+import androidx.compose.material.icons.filled.Menu
 import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
@@ -23,9 +26,49 @@ class MainActivity : ComponentActivity() {
 
         setContent {
             KotlinMusicTheme {
-                PlaylistScreen()
+                MainScreen()
             }
         }
+    }
+}
+
+@Composable
+fun MainScreen() {
+    val context = LocalContext.current
+    var selectedIndex by remember { mutableStateOf(0) }
+
+    Scaffold(
+        modifier = Modifier.fillMaxSize(),
+        bottomBar = {
+            BottomNavigation {
+                BottomNavigationItem(
+                    icon = { Icon(Icons.Default.Home, contentDescription = "Home") },
+                    label = { Text("Home") },
+                    selected = selectedIndex == 0,
+                    onClick = { selectedIndex = 0 }
+                )
+                BottomNavigationItem(
+                    icon = { Icon(Icons.Default.Menu, contentDescription = "Playlist") },
+                    label = { Text("Playlist") },
+                    selected = selectedIndex == 1,
+                    onClick = { selectedIndex = 1 }
+                )
+            }
+        }
+    ) { paddingValues ->
+        Column(modifier = Modifier.padding(paddingValues)) {
+            when (selectedIndex) {
+                0 -> HomePage()
+                1 -> PlaylistScreen()
+            }
+        }
+    }
+}
+
+@Composable
+fun HomePage() {
+    Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
+        Text(text = "Home Page")
     }
 }
 
@@ -34,24 +77,20 @@ fun PlaylistScreen() {
     val context = LocalContext.current
     val playlists = listOf("Playlist 1", "Playlist 2", "Playlist 3")
 
-    Scaffold(modifier = Modifier.fillMaxSize()) { paddingValues ->
-        Column(modifier = Modifier.padding(paddingValues)) {
-            LazyColumn(modifier = Modifier.fillMaxWidth()) {
-                items(playlists) { playlist ->
-                    Text(
-                        text = playlist,
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .clickable {
-                                val intent = Intent(context, DisplayMusicsActivity::class.java).apply {
-                                    putExtra("selected_playlist", playlist)
-                                }
-                                context.startActivity(intent)
-                            }
-                            .padding(16.dp)
-                    )
-                }
-            }
+    LazyColumn(modifier = Modifier.fillMaxWidth()) {
+        items(playlists) { playlist ->
+            Text(
+                text = playlist,
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .clickable {
+                        val intent = Intent(context, DisplayMusicsActivity::class.java).apply {
+                            putExtra("selected_playlist", playlist)
+                        }
+                        context.startActivity(intent)
+                    }
+                    .padding(16.dp)
+            )
         }
     }
 }
